Delete classes SettingsModuleSecureStorageService and LoginModuleSecureStorageService
