+++ MDFactory/Sources/Services/SecureStorage/SecureStore.swift
@@ -9,7 +9,6 @@ import Foundation
 
 class SecureStore {
 
-    //to be deleted after merge MDF-9
     /// Save user's password to secure storage
     /// - Parameters:
     ///   - userName: user's name
@@ -32,7 +31,6 @@ class SecureStore {
         }
     }
 
-    //to be deleted after merge MDF-9
     /// Read user's password from secure storage
     /// - Parameters:
     ///   - userName: user's name
@@ -64,7 +62,6 @@ class SecureStore {
         return password
     }
 
-    /// Delete user's password from secure storage
     /// - Parameters:
     ///   - userName: user's name
     static func deletePassword(userName: String) throws {
@@ -77,6 +74,78 @@ class SecureStore {
 
         guard status == errSecSuccess else {
             throw KeychainError.unexpectedStatus(status)
+        }
+    }
+
+    /// Update user's password to secure storage, if the one exists
+    /// - Parameters:
+    ///   - userName: user's name
+    ///   - newPassword: users's new password
+    static func updatePassword(userName: String, to newPassword: String) throws {
+        let query: [String: AnyObject] = [
+            kSecAttrAccount as String: userName as AnyObject,
+            kSecClass as String: kSecClassGenericPassword
+        ]
+
+        let attributes: [String: AnyObject] = [
+            kSecValueData as String: Data(newPassword.utf8) as AnyObject
+        ]
+
+        let status = SecItemUpdate(query as CFDictionary, attributes as CFDictionary)
+
+        guard status != errSecItemNotFound else {
+            throw KeychainError.itemNotFound
+        }
+
+        guard status == errSecSuccess else {
+            throw KeychainError.unexpectedStatus(status)
+        }
+    }
+
+    /// Check user's password existense in secure storage
+    /// - Parameters:
+    ///   - userName: user's name
+    static func isExists(userName: String) throws -> Bool {
+        let query: [String: AnyObject] = [
+            kSecAttrAccount as String: userName as AnyObject,
+            kSecClass as String: kSecClassGenericPassword
+        ]
+
+        let status = SecItemCopyMatching(query as CFDictionary, nil)
+
+        switch status {
+        case errSecSuccess, errSecInteractionNotAllowed:
+            return true
+        case errSecItemNotFound:
+            return false
+        default:
+            throw KeychainError.unexpectedStatus(status)
+        }
+    }
+}
+
+// MARK: - KeychainError
+
+enum KeychainError: Error {
+    case itemNotFound
+    case duplicateItem
+    case invalidItemFormat
+    case unexpectedStatus(OSStatus)
+}
+
+// MARK: - KeychainError Description
+
+extension KeychainError: LocalizedError {
+    var errorDescription: String? {
+        switch self {
+        case .itemNotFound:
+            return NSLocalizedString("Item not found", comment: "")
+        case .duplicateItem:
+            return NSLocalizedString("Trying to set existing item", comment: "")
+        case .invalidItemFormat:
+            return NSLocalizedString("Invalid item format", comment: "")
+        case .unexpectedStatus(let status):
+            return NSLocalizedString("Unexpected error \(status.description)", comment: "")
         }
     }
 }
